home controller - restructure files
