+++ src/api/legallead.permissions.api/Controllers/HomeController.cs
@@ -7,7 +7,7 @@ namespace legallead.permissions.api.Controllers
 {
     [Route("/")]
     [SuppressMessage("Sonar Qube Violation", "S6931:ASP.NET controller actions should not have a route template starting with \"/\"", Justification = "<Pending>")]
-    public class HomeController(
+    public partial class HomeController(
         IPaymentHtmlTranslator service,
         ISearchInfrastructure search,
         ISubscriptionInfrastructure subscriptionSvc,
@@ -27,208 +27,6 @@ namespace legallead.permissions.api.Controllers
             var content = IndexHtml;
             return Content(content, "text/html");
         }
-
-        [HttpGet("/payment-result")]
-        public async Task<IActionResult> PaymentLanding([FromQuery] string? sts, [FromQuery] string? id)
-        {
-            var isValid = await paymentSvc.IsRequestValid(sts, id);
-            var content =
-                isValid ? Properties.Resources.page_payment_completed
-                : Properties.Resources.page_payment_detail_invalid;
-            content = await paymentSvc.Transform(isValid, sts, id, content);
-            return Content(content, "text/html");
-        }
-
-        [HttpGet("/subscription-result")]
-        public async Task<IActionResult> UserLevelLanding([FromQuery] string? sts, [FromQuery] string? id)
-        {
-            var isValid = await paymentSvc.IsChangeUserLevel(sts, id);
-            var content =
-                isValid ? Properties.Resources.page_payment_completed
-                : Properties.Resources.page_level_request_completed;
-            content = await paymentSvc.TransformForPermissions(isValid, sts, id, content);
-            return Content(content, "text/html");
-        }
-
-        [HttpGet("/discount-result")]
-        public async Task<IActionResult> DiscountLanding([FromQuery] string? sts, [FromQuery] string? id)
-        {
-            var isValid = await paymentSvc.IsDiscountLevel(sts, id);
-            var content =
-                isValid ? Properties.Resources.page_payment_completed
-                : Properties.Resources.page_level_request_completed;
-            content = await paymentSvc.TransformForDiscounts(subscriptionSvc, isValid, id, content);
-            return Content(content, "text/html");
-        }
-
-        [HttpGet("/payment-checkout")]
-        [ServiceFilter(typeof(SearchPaymentCompleted))]
-        public async Task<IActionResult> PaymentCheckout([FromQuery] string? id)
-        {
-            var session = await paymentSvc.IsSessionValid(id);
-            if (session == null)
-            {
-                var nodata = Properties.Resources.page_payment_detail_invalid;
-                return Content(nodata, "text/html");
-            }
-            var ispaid = await paymentSvc.IsRequestPaid(session);
-            if (ispaid)
-            {
-                return await PaymentLanding("success", id);
-            }
-            var content = Properties.Resources.page_invoice_html;
-            content = paymentSvc.Transform(session, content);
-            return Content(content, "text/html");
-        }
-
-        [HttpGet("/subscription-checkout")]
-        public async Task<IActionResult> SubscriptionCheckout([FromQuery] string? id, string? sessionid)
-        {
-            var session = await paymentSvc.IsSubscriptionValid(id, sessionid);
-            if (session == null)
-            {
-                var nodata = Properties.Resources.page_payment_detail_invalid;
-                return Content(nodata, "text/html");
-            }
-            var ispaid = await paymentSvc.IsRequestPaid(session);
-            if (ispaid)
-            {
-                return await UserLevelLanding("success", id);
-            }
-            var clientSecret = await stripeSvc.FetchClientSecretValue(session);
-            var temp = Properties.Resources.page_invoice_subscription_html;
-            var content = paymentSvc.Transform(session, temp) ?? temp;
-            content = content.Replace("<!-- payment get intent index -->", clientSecret);
-            var result = Content(content, "text/html");
-            return result;
-        }
-
-
-
-        [HttpGet("/discount-checkout")]
-        public async Task<IActionResult> DiscountCheckout([FromQuery] string? id, string? sessionid)
-        {
-            var session = await paymentSvc.IsDiscountValid(id, sessionid);
-            if (session == null)
-            {
-                var nodata = Properties.Resources.page_payment_detail_invalid;
-                return Content(nodata, "text/html");
-            }
-            var ispaid = await paymentSvc.IsDiscountPaid(session);
-            if (ispaid)
-            {
-                // incorrect mapping need new landing for discount
-                return await DiscountLanding("success", id);
-            }
-            var content = Properties.Resources.page_invoice_discount_html;
-            var discountRequest = ModelMapper.Mapper.Map<DiscountRequestBo>(session);
-            content = paymentSvc.Transform(discountRequest, content);
-            return Content(content, "text/html");
-        }
-
-        [HttpPost("/payment-fetch-intent")]
-        public async Task<IActionResult> FetchIntent([FromBody] FetchIntentRequest request)
-        {
-            var session = await paymentSvc.IsSessionValid(request.Id);
-            if (session == null || string.IsNullOrEmpty(session.ClientId))
-            {
-                var nodata = Properties.Resources.page_payment_detail_invalid;
-                return Content(nodata, "text/html");
-            }
-            return Json(new { clientSecret = session.ClientId });
-        }
-
-
-        [HttpPost("/subscription-fetch-intent")]
-        public async Task<IActionResult> FetchSubscriptionIntent([FromBody] FetchIntentRequest request)
-        {
-            var nodata = Json(new { clientSecret = Guid.Empty.ToString("D") });
-            var session = await subscriptionSvc.GetLevelRequestById(request.Id, null);
-            if (session == null || string.IsNullOrEmpty(session.SessionId))
-            {
-                return nodata;
-            }
-            var clientSecret = await stripeSvc.FetchClientSecret(session);
-            return Json(new { clientSecret });
-        }
-
-        [HttpPost("/discount-fetch-intent")]
-        public async Task<IActionResult> FetchDiscountIntent([FromBody] FetchIntentRequest request)
-        {
-            var nodata = Json(new { clientSecret = Guid.Empty.ToString("D") });
-            var session = await subscriptionSvc.GetDiscountRequestById(request.Id, null);
-            if (session == null || string.IsNullOrEmpty(session.SessionId))
-            {
-                return nodata;
-            }
-            var clientSecret = await stripeSvc.FetchClientSecret(session);
-            return Json(new { clientSecret });
-        }
-
-        [Authorize]
-        [HttpPost("/payment-fetch-search")]
-        public async Task<IActionResult> FetchDownload([FromBody] FetchIntentRequest request)
-        {
-            var user = await infrastructure.GetUser(Request);
-            if (user == null || string.IsNullOrEmpty(user.Id)) return Unauthorized();
-            var isLocked = await _lockingDb.IsAccountLocked(user.Id);
-            if (isLocked)
-            {
-                return Forbid("Account is locked. Contact system administrator to unlock.");
-            }
-            var session = await paymentSvc.IsSessionValid(request.Id);
-            var ispaid = await paymentSvc.IsRequestPaid(session);
-            if (!ispaid)
-            {
-                return StatusCode(400, "Unable to find payment for associated download request.");
-            }
-            if (session == null || string.IsNullOrEmpty(session.JsText))
-            {
-                return StatusCode(400, "Unable to process request. One or more result artifacts are missing.");
-            }
-            var isdownload = await paymentSvc.IsRequestDownloadedAndPaid(session);
-            if (isdownload)
-            {
-                return StatusCode(400, "Associated download result has already been delivered.");
-            }
-            var dwnload = await paymentSvc.GetDownload(session);
-            return Ok(dwnload);
-        }
-
-        [Authorize]
-        [HttpPost("/rollback-download")]
-        public async Task<IActionResult> RollbackDownload([FromBody] DownloadResetRequest request)
-        {
-            var user = await infrastructure.GetUser(Request);
-            if (user == null || !user.UserName.Equals(request.UserId, StringComparison.OrdinalIgnoreCase))
-                return Unauthorized();
-            var isLocked = await _lockingDb.IsAccountLocked(user.Id);
-            if (isLocked)
-            {
-                return Forbid("Account is locked. Contact system administrator to unlock.");
-            }
-            var session = await paymentSvc.IsSessionValid(request.ExternalId);
-            var ispaid = await paymentSvc.IsRequestPaid(session);
-            if (!ispaid)
-            {
-                return StatusCode(400, "Unable to find payment for associated download request.");
-            }
-            if (session == null || string.IsNullOrEmpty(session.JsText))
-            {
-                return StatusCode(400, "Unable to process request. One or more result artifacts are missing.");
-            }
-            var isdownload = await paymentSvc.IsRequestDownloadedAndPaid(session);
-            if (!isdownload)
-            {
-                return StatusCode(400, "Associated download result has not been downloaded.");
-            }
-            request.UserId = user.Id;
-            request.ExternalId = session.InvoiceId;
-            var dwnload = await paymentSvc.ResetDownload(request);
-            if (dwnload == null) return UnprocessableEntity("Unable to perform reset. Process rejected by server.");
-            return Ok(dwnload);
-        }
-
         private static string? _index;
         private static string IndexHtml => _index ??= GetIndex();
